deleted LoginActivity and a duplicate LoginScreen
